Added avalonia implementation of the library, moved Wpf implementation to .Wpf
