+++ .gitignore
@@ -34,6 +34,8 @@ bld/
 
 # Visual Studio 2015/2017 cache/options directory
 .vs/
+# Rider cache directory
+.idea/
 # Uncomment if you have tasks that create the project's static files in wwwroot
 #wwwroot/
 
@@ -358,3 +360,5 @@ MigrationBackup/
 
 # Fody - auto-generated XML schema
 FodyWeavers.xsd
+ProgressDialog/.DS_Store
+.DS_Store
