+++ ProgressDialog/ProgressDialog.Avalonia.Example/App.axaml
@@ -1,0 +1,12 @@
+<Application xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             x:Class="ProgressDialog.Avalonia.Example.App"
+             xmlns:local="using:ProgressDialog.Avalonia.Example"
+             RequestedThemeVariant="Default">
+             <!-- "Default" ThemeVariant follows system theme variant. "Dark" or "Light" are other available options. -->
+
+  
+    <Application.Styles>
+        <FluentTheme />
+    </Application.Styles>
+</Application>
