+++ ProgressDialog/ProgressDialog.Avalonia.Example/App.axaml.cs
@@ -1,0 +1,24 @@
+using Avalonia;
+using Avalonia.Controls.ApplicationLifetimes;
+using Avalonia.Markup.Xaml;
+
+namespace ProgressDialog.Avalonia.Example;
+
+public partial class App : Application
+{
+    public override void Initialize()
+    {
+        AvaloniaXamlLoader.Load(this);
+    }
+
+    public override void OnFrameworkInitializationCompleted()
+    {
+        if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
+        {
+            desktop.MainWindow = new MainWindow();
+            desktop.MainWindow.DataContext = new ProgressDialogExampleViewModel(desktop.MainWindow);
+        }
+
+        base.OnFrameworkInitializationCompleted();
+    }
+}
