+++ ProgressDialog/ProgressDialog.Avalonia.Example/MainWindow.axaml.cs
@@ -1,0 +1,13 @@
+using Avalonia.Controls;
+using ProgressDialog.Avalonia.Example;
+
+namespace ProgressDialog.Avalonia.Example;
+
+public partial class MainWindow : Window
+{
+    public MainWindow()
+    {
+        DataContext = new ProgressDialogExampleViewModel(this);
+        InitializeComponent();
+    }
+}
