+++ ProgressDialog/ProgressDialog.Avalonia/ProgressDialogUserControl.axaml
@@ -1,0 +1,55 @@
+<!--
+____ Copyright Start ____
+
+MIT License
+
+Copyright (c) 2020 ax-meyer
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+____ Copyright End ____
+-->
+
+<controls:UserControl x:Class="ProgressDialog.Avalonia.ProgressDialogUserControl"
+                      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                      xmlns:local="clr-namespace:ProgressDialog.Avalonia"
+                      xmlns:controls="https://github.com/avaloniaui"
+                      xmlns:progressDialog="clr-namespace:ProgressDialog;assembly=ProgressDialogStatus"
+                      mc:Ignorable="d"
+                      d:DesignHeight="450" d:DesignWidth="800"
+                      x:CompileBindings="False"
+                      >
+    <controls:Grid>
+        <controls:Grid.RowDefinitions>
+            <controls:RowDefinition />
+            <controls:RowDefinition />
+            <controls:RowDefinition />
+        </controls:Grid.RowDefinitions>
+        <controls:TextBlock Text="{controls:Binding Message, Mode=OneWay}" controls:Grid.Row="0"
+                            TextWrapping="Wrap" />
+        <controls:Grid controls:Grid.Row="1">
+            <controls:ProgressBar Value="{controls:Binding ProgressPercent, Mode=OneWay}" />
+            <controls:TextBlock HorizontalAlignment="Center" VerticalAlignment="Center"
+                                Text="{controls:Binding ProgressPercent, StringFormat={}{0}%}" />
+        </controls:Grid>
+        <controls:Button Content="Cancel" Command="{controls:Binding CancelCommand}" controls:Grid.Row="2" />
+    </controls:Grid>
+</controls:UserControl>
