+++ ProgressDialog/ProgressDialog.Wpf.Example/App.xaml.cs
@@ -28,30 +28,17 @@ ____ Copyright End ____
 
 ﻿using System;
 using System.Collections.Generic;
+using System.Configuration;
+using System.Data;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
-namespace ProgressDialogExample
+namespace ProgressDialog.Wpf.Example;
+
+/// <summary>
+/// Interaction logic for App.xaml
+/// </summary>
+public partial class App : Application
 {
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : Window
-    {
-        public MainWindow()
-        {
-            DataContext = new ProgressDialogExampleViewModel(this);
-            InitializeComponent();
-        }
-    }
 }
