+++ ProgressDialog/ProgressDialog.Wpf.Example/AssemblyInfo.cs
