+++ ProgressDialog/ProgressDialog.Wpf/AssemblyInfo.cs
