+++ ProgressDialog/ProgressDialog.Wpf/ProgressDialog.Wpf.csproj
@@ -1,0 +1,30 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <LangVersion>latest</LangVersion>
+        <Nullable>enable</Nullable>
+        <WarningsAsErrors>Nullable</WarningsAsErrors>
+        <EnableWindowsTargeting>true</EnableWindowsTargeting>
+        <UseWPF>true</UseWPF>
+
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <Authors>Alexander Meyer</Authors>
+        <Company />
+        <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+        <RepositoryUrl>https://github.com/ax-meyer/ProgressDialog/</RepositoryUrl>
+        <PackageProjectUrl>https://github.com/ax-meyer/ProgressDialog/</PackageProjectUrl>
+        <PackageTags>WPF, MVVM, Progress, Dialog, Status</PackageTags>
+        <PackageOutputPath>../../</PackageOutputPath>
+        <Description>An easy-to-integrate progess indicator for MVVM pattern based WPF applications. Features a status text &amp; progress bar that can be updated from long running tasks in the model through a IProgressStatus interface.
+
+            The interface is also provided seperateley in the ProgressDialogStatus package, which is a netStandard 2.0 package without dependency on WPF, so it can be integrated into netStandard libraries. It also contains a ready-to-use sample implementation of the interface.
+        </Description>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\ProgressDialogStatus\ProgressDialogStatus.csproj" />
+    </ItemGroup>
+</Project>
