+++ ProgressDialog/ProgressDialog.Wpf/ProgressDialogUserControl.xaml
@@ -26,21 +26,28 @@ SOFTWARE.
 ____ Copyright End ____
 -->
 
-<Window x:Class="ProgressDialog.ProgressDialogWindow"
+<UserControl x:Class="ProgressDialog.Wpf.ProgressDialogUserControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-        x:Name="ProgressDialogWin"
-             xmlns:local="clr-namespace:ProgressDialog"
+             xmlns:local="clr-namespace:ProgressDialog.Wpf"
              mc:Ignorable="d" 
-        Width="300"
-        MinWidth="100"
-        Height="150"
-        WindowStartupLocation="CenterOwner"
-        >
-    
+             d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
-        <local:ProgressDialogUserControl />
+        <Grid.RowDefinitions>
+            <RowDefinition/>
+            <RowDefinition/>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition/>
+        </Grid.ColumnDefinitions>
+        <TextBlock Text="{Binding Message, Mode=OneWay}" Grid.Row="0" TextWrapping="Wrap"/>
+        <Grid Grid.Row="1">
+            <ProgressBar Value="{Binding ProgressPercent, Mode=OneWay}"/>
+            <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center" Text="{Binding ProgressPercent, StringFormat={}{0}%}"/>
+        </Grid>
+        <Button Content="Cancel" Command="{Binding CancelCommand}" Grid.Row="2"/>
     </Grid>
-</Window>
+</UserControl>
