+++ ProgressDialog/ProgressDialog.Wpf/ProgressDialogWindow.xaml
@@ -26,28 +26,21 @@ SOFTWARE.
 ____ Copyright End ____
 -->
 
-<UserControl x:Class="ProgressDialog.ProgressDialogUserControl"
+<Window x:Class="ProgressDialog.Wpf.ProgressDialogWindow"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:ProgressDialog"
+        x:Name="ProgressDialogWin"
+             xmlns:local="clr-namespace:ProgressDialog.Wpf"
              mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
+        Width="300"
+        MinWidth="100"
+        Height="150"
+        WindowStartupLocation="CenterOwner"
+        >
+    
     <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition/>
-            <RowDefinition/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-        <TextBlock Text="{Binding Message, Mode=OneWay}" Grid.Row="0" TextWrapping="Wrap"/>
-        <Grid Grid.Row="1">
-            <ProgressBar Value="{Binding ProgressPercent, Mode=OneWay}"/>
-            <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center" Text="{Binding ProgressPercent, StringFormat={}{0}%}"/>
-        </Grid>
-        <Button Content="Cancel" Command="{Binding CancelCommand}" Grid.Row="2"/>
+        <local:ProgressDialogUserControl />
     </Grid>
-</UserControl>
+</Window>
