--- ProgressDialog/ProgressDialog/ProgressDialog.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
-
-  <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1</TargetFrameworks>
-    <UseWPF>true</UseWPF>
-    <Authors>Alexander Meyer</Authors>
-    <Company />
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <RepositoryUrl>https://github.com/ax-meyer/ProgressDialog/</RepositoryUrl>
-  </PropertyGroup>
-  
-  <PropertyGroup>
-    <Version>1.2.4</Version>
-    <PackageProjectUrl>https://github.com/ax-meyer/ProgressDialog/</PackageProjectUrl>
-    <PackageTags>WPF, MVVM, Progress, Dialog, Status</PackageTags>
-    <PackageOutputPath>../</PackageOutputPath>
-    <Description>An easy-to-integrate progess indicator for MVVM pattern based WPF applications. Features a status text &amp; progress bar that can be updated from long running tasks in the model through a IProgressStatus interface. 
-
-The interface is also provided seperateley in the ProgressDialogStatus package, which is a netStandard 2.0 package without dependency on WPF, so it can be integrated into netStandard libraries. It also contains a ready-to-use sample implementation of the interface.
-</Description>
-    <AssemblyVersion>1.2.4.0</AssemblyVersion>
-    <FileVersion>1.2.4.0</FileVersion>
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <PackageReference Include="ProgressDialogStatus" Version="1.2.3" />
-  </ItemGroup>
-
-</Project>
