--- ProgressDialog/ProgressDialogExample/ProgressDialogExample.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
-
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net7.0-windows</TargetFramework>
-    <UseWPF>true</UseWPF>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Prism.Core" Version="7.2.0.1422" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ProgressDialog\ProgressDialog.csproj" />
-  </ItemGroup>
-
-</Project>
